refactor(components): change name and add priority return
